ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-openai==0.2.6
python-dotenv==1.0.1
redis==5.2.1
